+++ CHANGELOG.md
@@ -13,6 +13,7 @@ how this file is formatted and how the process works.
 - `waytostay#quote` and `waytostay#booking`, which calls the above client and respond accordingly
 - `byebug` for test/development debugging
 - `Concierge::SafeAccessHash#missing_keys_from?` utils method
+- `SyncProcess` entity and database, recording every synchronisation process run on Concierge.
 
 ### Changed
 - HTTPStubbing supports matching by body and headers, by including `strict: true` option to `stub_call`